[Core] [Metadata] Reconstruct metadata synchronization method
